- Add transaction table view data source class - Add logic to transaction VC to update UI, wire up datasource, and ability to add transactions - Add delegate protocol to Create transaction VC and code to create transaction object

+++ Bank Account App/Bank Account App/Views/NewTransactionView.swift
@@ -21,7 +21,7 @@ class NewTransactionView:UIView
     fileprivate var selectedIndex:Int = 0
     
     // MARK: Public Computed Properties
-    var amount:Float?
+    var amount:Double?
     {
         get
         {
@@ -31,7 +31,7 @@ class NewTransactionView:UIView
                 return nil
             }
             
-            return Float(text)
+            return Double(text)
         }
     }
